rename providers module to parsers
