+++ docs/api_reference/document_search/processing.md
@@ -3,22 +3,22 @@
 ::: ragbits.document_search.ingestion.document_processor.DocumentProcessorRouter
 
 ## Providers
-::: ragbits.document_search.ingestion.providers.base.BaseProvider
+::: ragbits.document_search.ingestion.parsers.base.BaseProvider
     options:
       heading_level: 3
 
-::: ragbits.document_search.ingestion.providers.dummy.DummyProvider
+::: ragbits.document_search.ingestion.parsers.dummy.DummyProvider
     options:
       heading_level: 3
 
-::: ragbits.document_search.ingestion.providers.unstructured.UnstructuredDefaultProvider
+::: ragbits.document_search.ingestion.parsers.unstructured.UnstructuredDefaultProvider
     options:
       heading_level: 3
 
-::: ragbits.document_search.ingestion.providers.unstructured.UnstructuredImageProvider
+::: ragbits.document_search.ingestion.parsers.unstructured.UnstructuredImageProvider
     options:
       heading_level: 3
 
-::: ragbits.document_search.ingestion.providers.unstructured.UnstructuredPdfProvider
+::: ragbits.document_search.ingestion.parsers.unstructured.UnstructuredPdfProvider
     options:
       heading_level: 3
