+++ examples/document-search/multimodal_qdrant.py
@@ -42,7 +42,7 @@ from ragbits.document_search import DocumentSearch
 from ragbits.document_search.documents.document import DocumentMeta, DocumentType
 from ragbits.document_search.documents.sources import LocalFileSource
 from ragbits.document_search.ingestion.document_processor import DocumentProcessorRouter
-from ragbits.document_search.ingestion.providers.dummy import DummyImageProvider
+from ragbits.document_search.ingestion.parsers.dummy import DummyImageProvider
 
 IMAGES_PATH = Path(__file__).parent / "images"
