+++ examples/evaluation/document-search/basic/optimize.py
@@ -55,7 +55,7 @@ config = {
                 },
                 "providers": {
                     "txt": {
-                        "type": "ragbits.document_search.ingestion.providers.unstructured:UnstructuredDefaultProvider",
+                        "type": "ragbits.document_search.ingestion.parsers.unstructured:UnstructuredDefaultProvider",
                     },
                 },
                 "source": {
